# api/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box WebUI Local API
  version: 0.1.0
  description: Local loopback API for the sing-box WebUI control plane.
servers:
  - url: http://127.0.0.1:11872
paths:
...
  /api/v1/status:
    get:
      operationId: getStatus
      responses:
        '200':
          description: Current control-plane component status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Authentication succeeded and an HttpOnly session cookie was issued.
        default:
          $ref: '#/components/responses/Error'
  /api/v1/auth/logout:
    post:
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: The browser session cookie was cleared.
        default:
          $ref: '#/components/responses/Error'
  /api/v1/traffic-policy:
    get:
      operationId: getTrafficPolicy
      responses:
        '200':
          description: Persisted automatic download policy and its current runtime state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficPolicy'
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: updateTrafficPolicy
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrafficPolicy'
      responses:
        '200':
          description: Updated policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficPolicy'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/dns/profile:
    get:
      operationId: getDnsProfile
      responses:
        '200':
          description: Persisted DNS profile used for TUN-mode name resolution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnsProfile'
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: updateDnsProfile
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDnsProfile'
      responses:
        '200':
          description: Updated DNS profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnsProfile'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/events:
    get:
      operationId: streamEvents
      responses:
        '200':
          description: Server-sent event stream. The first event is a snapshot.
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/subscriptions:
    get:
      operationId: listSubscriptions
      responses:
        '200':
          description: Stored subscriptions without credential material.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
    post:
      operationId: createSubscription
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscription'
      responses:
        '201':
          description: Subscription stored and initial fetch attempted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/order:
    put:
      operationId: reorderSubscriptions
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionOrder'
      responses:
        '200':
          description: Subscriptions in their persisted display order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
    get:
      operationId: getSubscription
      responses:
        '200':
          description: Subscription and sanitized node list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
    patch:
      operationId: updateSubscription
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscription'
      responses:
        '200':
          description: Updated subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteSubscription
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Subscription removed.
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}/refresh:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    post:
      operationId: refreshSubscription
      responses:
        '200':
          description: Refreshed subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}/activate:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    post:
      operationId: activateSubscription
      responses:
        '200':
          description: Active subscription changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}/selection:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    put:
      operationId: selectNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [nodeId]
              properties:
                nodeId:
                  type: string
      responses:
        '200':
          description: Selected node changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/subscriptions/{id}/latency:
    parameters:
      - $ref: '#/components/parameters/SubscriptionID'
      - $ref: '#/components/parameters/CSRFToken'
    post:
      operationId: testNodeLatency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LatencyRequest'
      responses:
        '200':
          description: End-to-end HTTPS delay results measured through temporary sing-box outbounds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatencyResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/pools:
    get:
      operationId: listNodePools
      responses:
        '200':
          description: Stored cross-subscription node pools with live member resolution.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NodePool'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createNodePool
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodePool'
      responses:
        '201':
          description: Node pool created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePool'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/pools/order:
    put:
      operationId: reorderNodePools
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodePoolOrder'
      responses:
        '200':
          description: Node pools in their persisted display order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NodePool'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/pools/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getNodePool
      responses:
        '200':
          description: Node pool details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePool'
        default:
          $ref: '#/components/responses/Error'
    patch:
      operationId: updateNodePool
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNodePool'
      responses:
        '200':
          description: Node pool updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePool'
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteNodePool
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Node pool deleted.
        default:
          $ref: '#/components/responses/Error'
  /api/v1/runtime:
    get:
      operationId: getRuntime
      responses:
        '200':
          description: Current proxy runtime and capabilities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runtime'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/links:
    get:
      operationId: listLinks
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Case-insensitive substring match across host, URL/domain, node, network, type, and chain.
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter to live (true) or closed (false) connections.
        - name: sort
          in: query
          schema:
            type: string
          description: >-
            Comma-separated sort keys applied in order. Prefix with - for descending.
            Keys: host, url, node, upload, download, uploadRate, downloadRate, startedAt.
            Active connections always sort before closed ones.
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 1000
      responses:
        '200':
          description: Snapshot of observed proxy connections.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkSnapshot'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/links/clear:
    post:
      operationId: clearLinks
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: Connection cache cleared. Monitoring continues.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status:
                    type: string
        default:
          $ref: '#/components/responses/Error'
  /api/v1/rules:
    get:
      operationId: listRules
      responses:
        '200':
          description: Manual, subscription, and built-in routing rules in evaluation order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Rule'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createRule
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRule'
      responses:
        '201':
          description: Manual rule created and reloaded when the proxy is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/rules/order:
    put:
      operationId: reorderRules
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleOrder'
      responses:
        '200':
          description: Manual rules reordered and reloaded when the proxy is running.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Rule'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/rules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateRule
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRule'
      responses:
        '200':
          description: Rule updated and reloaded when the proxy is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteRule
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Manual rule deleted and reloaded when the proxy is running.
        default:
          $ref: '#/components/responses/Error'
  /api/v1/rule-pools:
    get:
      operationId: listRulePools
      responses:
        '200':
          description: Rule pools in evaluation order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RulePool'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createRulePool
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRulePool'
      responses:
        '201':
          description: Rule pool created and reloaded when the proxy is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulePool'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/rule-pools/order:
    put:
      operationId: reorderRulePools
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RulePoolOrder'
      responses:
        '200':
          description: Rule pools reordered and reloaded when the proxy is running.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RulePool'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/rule-pools/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateRulePool
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRulePool'
      responses:
        '200':
          description: Rule pool updated atomically and reloaded when the proxy is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulePool'
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteRulePool
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Rule pool deleted and reloaded when the proxy is running.
        default:
          $ref: '#/components/responses/Error'
  /api/v1/runtime/apply:
    post:
      operationId: applyRuntime
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyRuntime'
      responses:
        '200':
          description: Selected node or node pool and proxy mode applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runtime'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/runtime/stop:
    post:
      operationId: stopRuntime
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: Proxy stopped and system proxy state restored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runtime'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/connectivity:
    get:
      operationId: listConnectivityTargets
      responses:
        '200':
          description: Stored quick connectivity test targets.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConnectivityTarget'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createConnectivityTarget
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectivityTarget'
      responses:
        '201':
          description: Created connectivity target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectivityTarget'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/connectivity/test:
    post:
      operationId: testAllConnectivity
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: Direct and proxied latency for every stored target, measured concurrently.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectivityTestResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/connectivity/diagnostic:
    post:
      operationId: runConnectivityDiagnostic
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectivityDiagnosticInput'
      responses:
        '200':
          description: Quality or exit-IP details measured through the active proxy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectivityDiagnosticResult'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/connectivity/{id}/test:
    parameters:
      - $ref: '#/components/parameters/ConnectivityTargetID'
    post:
      operationId: testConnectivity
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: Direct and proxied latency for a single target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectivityTestResponse'
        default:
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    SubscriptionID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Status:
      type: object
      additionalProperties: false
      required: [service, version, components, timestamp]
      properties:
        service:
          type: string
        version:
          type: string
        components:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ComponentStatus'
        timestamp:
          type: string
          format: date-time
...
    UpdateTrafficPolicy:
      type: object
      additionalProperties: false
      required: [enabled, downloadPoolId, triggerRateBytesPerSecond, triggerDurationSeconds, releaseRateBytesPerSecond, releaseDurationSeconds, cooldownSeconds]
      properties:
        enabled:
          type: boolean
        downloadPoolId:
          type: string
        triggerRateBytesPerSecond:
          type: integer
          format: int64
          minimum: 65536
        triggerDurationSeconds:
          type: integer
          minimum: 2
          maximum: 300
        releaseRateBytesPerSecond:
          type: integer
          format: int64
          minimum: 0
        releaseDurationSeconds:
          type: integer
          minimum: 5
          maximum: 3600
        cooldownSeconds:
          type: integer
          minimum: 0
          maximum: 86400
...